server/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
alembic==1.10.4
pydantic==1.10.7
psycopg2-binary==2.9.6  # PostgreSQL driver
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0
httpx==0.24.0
openai==0.27.6
python-dotenv==1.0.0 